+++ packages/component/src/block/accountStatus/desktop/assets.tsx
@@ -54,7 +54,7 @@ export const Assets: FC<AssetsProps> = (props) => {
       return false;
     }
 
-    const isTestnetChain = parseInt(chainId, 16) === 421613;
+    const isTestnetChain = parseInt(chainId) === 421613;
 
     return state.status === AccountStatusEnum.EnableTrading && isTestnetChain;
   }, [state.status, connectedChain]);