+++ packages/component/src/layout/types.ts
@@ -1,0 +1,12 @@
+import { HTMLAttributes, ReactNode } from "react";
+
+// export type BreakPoint = "xs" | "sm" | "md" | "lg" | "xl" | "xxl";
+export type BreakPoint = "mobile" | "tablet";
+
+export interface LayoutBaseProps extends HTMLAttributes<HTMLDivElement> {
+  //   breakpoint?: BreakPoint;
+  tablet?: ReactNode;
+  mobile?: ReactNode;
+
+  asChild?: boolean;
+}