+++ packages/component/src/page/trading/context/tradingPageContext.tsx
@@ -1,13 +1,13 @@
 import { API } from "@orderly.network/types";
-import React, { PropsWithChildren, useMemo } from "react";
-import { usePrivateQuery } from "@orderly.network/hooks";
+import React, { PropsWithChildren, ReactNode, useContext } from "react";
+import { TradingFeatures } from "../features";
 
 export interface TradingPageContextValue {
   onSymbolChange?: (symbol: API.Symbol) => void;
   symbol: string;
 
-  // positionCount: number;
-  // pendingCount: number;
+  disableFeatures: TradingFeatures[];
+  overrides?: Record<TradingFeatures, ReactNode>;
 }
 
 export const TradingPageContext = React.createContext<TradingPageContextValue>(
@@ -17,13 +17,21 @@ export const TradingPageContext = React.createContext<TradingPageContextValue>(
 export interface TradingPageProviderProps {
   onSymbolChange?: (symbol: API.Symbol) => void;
   symbol: string;
+  disableFeatures?: TradingFeatures[];
+  overrides?: Record<TradingFeatures, ReactNode>;
 }
+
+export const useTradingPageContext = () => {
+  return useContext(TradingPageContext);
+};
 
 export const TradingPageProvider: React.FC<
   PropsWithChildren<TradingPageProviderProps>
-> = ({ children, onSymbolChange, symbol }) => {
+> = ({ children, onSymbolChange, symbol, disableFeatures = [], overrides }) => {
   return (
-    <TradingPageContext.Provider value={{ onSymbolChange, symbol }}>
+    <TradingPageContext.Provider
+      value={{ onSymbolChange, symbol, disableFeatures, overrides }}
+    >
       {children}
     </TradingPageContext.Provider>
   );