+++ packages/types/src/constants.ts
@@ -25,6 +25,6 @@ export const ARBITRUM_MAINNET_CHAINID = 42161;
 export const ARBITRUM_MAINNET_CHAINID_HEX = "0xa4b1";
 export const ARBITRUM_TESTNET_CHAINID_HEX = "0x66EED";
 
-export const MEDIA_TABLE = "(max-width: 768px)";
+export const MEDIA_TABLET = "(max-width: 768px)";
 
 export const DEPOSIT_FEE_RATE = 1.05;