+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,61 @@
 # examples
 
+## 0.1.92-alpha.7
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/react@0.1.31-alpha.7
+  - @orderly.network/web3-onboard@0.0.28-alpha.7
+  - @orderly.network/hooks@1.0.28-alpha.7
+
+## 0.1.92-alpha.6
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/react@0.1.31-alpha.6
+  - @orderly.network/web3-onboard@0.0.28-alpha.6
+  - @orderly.network/hooks@1.0.28-alpha.6
+
+## 0.1.92-alpha.5
+
+### Patch Changes
+
+- fix bug
+- fix bug
+- fix bug
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/react@0.1.31-alpha.5
+  - @orderly.network/web3-onboard@0.0.28-alpha.5
+  - @orderly.network/hooks@1.0.28-alpha.5
+
+## 0.1.92-alpha.4
+
+### Patch Changes
+
+- fix bugs
+- fix bugs
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/react@0.1.31-alpha.4
+  - @orderly.network/web3-onboard@0.0.28-alpha.4
+  - @orderly.network/hooks@1.0.28-alpha.4
+
+## 0.1.92-alpha.3
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/react@0.1.31-alpha.3
+  - @orderly.network/web3-onboard@0.0.28-alpha.3
+  - @orderly.network/hooks@1.0.28-alpha.3
+
 ## 0.1.92-alpha.2
 
 ### Patch Changes