+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @orderly.network/futures
 
+## 1.0.27-alpha.7
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.27-alpha.7
+  - @orderly.network/utils@0.0.79-alpha.7
+
+## 1.0.27-alpha.6
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.27-alpha.6
+  - @orderly.network/utils@0.0.79-alpha.6
+
+## 1.0.27-alpha.5
+
+### Patch Changes
+
+- fix bug
+- fix bug
+- fix bug
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/types@0.1.27-alpha.5
+  - @orderly.network/utils@0.0.79-alpha.5
+
+## 1.0.27-alpha.4
+
+### Patch Changes
+
+- fix bugs
+- fix bugs
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/types@0.1.27-alpha.4
+  - @orderly.network/utils@0.0.79-alpha.4
+
+## 1.0.27-alpha.3
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/types@0.1.27-alpha.3
+  - @orderly.network/utils@0.0.79-alpha.3
+
 ## 1.0.27-alpha.2
 
 ### Patch Changes