+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,38 @@
 # tsconfig
 
+## 0.0.87-alpha.7
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.87-alpha.6
+
+### Patch Changes
+
+- fix bug
+
+## 0.0.87-alpha.5
+
+### Patch Changes
+
+- fix bug
+- fix bug
+- fix bug
+
+## 0.0.87-alpha.4
+
+### Patch Changes
+
+- fix bugs
+- fix bugs
+
+## 0.0.87-alpha.3
+
+### Patch Changes
+
+- fix bug
+
 ## 0.0.87-alpha.2
 
 ### Patch Changes