+++ packages/types/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @orderly.network/types
 
+## 0.1.27-alpha.7
+
+### Patch Changes
+
+- fix bug
+
+## 0.1.27-alpha.6
+
+### Patch Changes
+
+- fix bug
+
+## 0.1.27-alpha.5
+
+### Patch Changes
+
+- fix bug
+- fix bug
+- fix bug
+
+## 0.1.27-alpha.4
+
+### Patch Changes
+
+- fix bugs
+- fix bugs
+
+## 0.1.27-alpha.3
+
+### Patch Changes
+
+- fix bug
+
 ## 0.1.27-alpha.2
 
 ### Patch Changes